Move class GraphCache to own file
